chore: 공통적으로 사용하는 바인딩 어댑터 파일 분리
