+++ app/src/main/java/woowacourse/shopping/presentation/common/GlideBindingAdapters.kt
@@ -1,0 +1,14 @@
+package woowacourse.shopping.presentation.common
+
+import android.widget.ImageView
+import androidx.databinding.BindingAdapter
+import com.bumptech.glide.Glide
+
+@BindingAdapter("urlToImage")
+fun ImageView.bindUrlToImage(imageUrl: String?) {
+    imageUrl?.let { url ->
+        Glide.with(context)
+            .load(url)
+            .into(this)
+    }
+}
